+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project are documented in this file.
 
+## Unreleased
+
+- Added: `payload` and `method` arguments.
+
 ## 0.1.0
 
 - Added: `use-api` hook.